refactor: enum 패키지 분리
