+++ src/main/java/com/sparta/bapzip/menu/domain/entity/MenuEntity.java
@@ -1,6 +1,7 @@
 package com.sparta.bapzip.menu.domain.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.sparta.bapzip.menu.domain.enums.MenuStatus;
 import com.sparta.bapzip.ordermenu.domain.entity.OrderMenuEntity;
 import com.sparta.bapzip.shop.domain.entity.ShopEntity;
 import jakarta.persistence.*;
