+++ src/main/java/com/sparta/bapzip/menu/domain/enums/MenuStatus.java
@@ -1,4 +1,4 @@
-package com.sparta.bapzip.menu.domain.entity;
+package com.sparta.bapzip.menu.domain.enums;
 
 public enum MenuStatus {
     AVAILABLE,   // 판매중
